+++ phpstan.neon
@@ -1,6 +1,5 @@
 parameters:
     paths:
-        - bin
         - src
         - tests
